Put emphasize on some stuff
Improve the link between the general command and the stuff you can have inside.

I wonder if we should remove all the underscores from around the exemples. After all, if the idea is that you replace AESTHETIC by "fill" or "color" in the exemple comand, then you don't add extra underscores. But mostly, it's a lot of weird characters and it makes the exempe so much harder to read...